refactor: getTimeDiff함수 분리
