+++ public/mocks/detailCommunity.json
@@ -22,7 +22,7 @@
     "comments": [
       {
         "id": "comments id : 255617b0-446c-477e-96c9-cb9fe54bc6cc",
-        "nickname": "아아무개",
+        "nickname": "이아무개",
         "comment": "댓글을 달고있는 나는 멋져",
         "date": "2023-09-16",
         "userId": "255617b0-446c-477e-96c9-cb9fe54bc6cc"
