+++ src/components/Community/CommunityDetail/ContentsInfo/ContentInfo.tsx
@@ -1,17 +1,20 @@
+import { getTimeDiffString } from 'components/Community/utils/getTimeDiffString';
+
 import * as S from './ContentInfo.styles';
 
 type Props = {
   profileImage?: string;
   nickname?: string;
+  date?: string;
 };
 
-export const ContentInfo = ({ profileImage, nickname }: Props) => {
+export const ContentInfo = ({ profileImage, nickname, date }: Props) => {
   return (
     <S.ContentInfo>
       <S.Profile src={profileImage} />
       <div>
         <S.Nickname>{nickname}님</S.Nickname>
-        <S.TimeDiff>12잔</S.TimeDiff>
+        <S.TimeDiff>{getTimeDiffString(date as string)}</S.TimeDiff>
       </div>
     </S.ContentInfo>
   );
